% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{base_final_tidy}
\alias{base_final_tidy}
\title{base_final_tidy}
\format{
um data frame com 193 linhas e 17 variaveis:

\describe{
\item{num_bo }{numero do bo relacionado a infracao.}
\item{ano_bo}{ano da infracao.}
\item{numero_boletim}{numero do boletim.}
\item{delegacia}{local que foi registrada a infracao.}
\item{natureza_vinculada}{natureza juridica da infracao.}
\item{infracoes}{tipificacao.}
\item{bo_iniciado}{data e hora que o bo foi iniciado.}
\item{bo_emitido}{data e hora que o bo foi emitido.}
\item{dhocorrencia}{data e hora da ocorrencia.}
\item{datacomunicacao}{data da comunicacao do crime.}
\item{endereco}{endereco da infracao.}
\item{numero}{numero do endereco da infracao.}
\item{uf}{uf do endereco da infracao.}
\item{latitude}{latitude do endereco da infracao.}
\item{longitude}{longitude do endereco da infracao.}
\item{descricaolocal}{descricao do local da infracao.}
}
}
\source{
base_final_tidy
}
\usage{
base_final_tidy
}
\description{
Um conjunto de dados que contem informacoes os genericas sobre os casos.
}
\keyword{datasets}
