% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_importacao.R
\name{ler_uma_base}
\alias{ler_uma_base}
\title{Função de leitura das bases}
\usage{
ler_uma_base(path)
}
\arguments{
\item{path}{caminho para a pasta que esta as bases.}
}
\description{
Essa função facilita a leitura das bases principais
}
