% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa_string.R
\name{limpa_strings}
\alias{limpa_strings}
\title{Função de limpar os nomes}
\usage{
limpa_strings(str)
}
\arguments{
\item{str}{string que esta suja.}
}
\description{
Essa função facilita a limpeza de strings
}
