% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{partes_tidy}
\alias{partes_tidy}
\title{partes_tidy}
\format{
um data frame com 267 linhas e 20 variaveis:

\describe{
\item{num_bo }{numero do bo relacionado a infracao.}
\item{ano_bo}{ano da infracao.}
\item{bo_autoria}{identificacao do autor do crime.}
\item{flagrante}{informacao de flagrante.}
\item{exame}{local do exame.}
\item{solucao}{solucao do bo.}
\item{status}{informacao de consumado ou tentado.}
\item{tipopessoa}{informacacao de vitima ou autor.}
\item{vitimafatal}{informacao de vitima fatal ou nao.}
\item{naturalidade}{naturalidade da vitima.}
\item{nacionalidade}{nacionalidade da vitima.}
\item{sexo}{sexo da vitima.}
\item{datanascimento}{data nascimento da vitima.}
\item{idade}{idade da vitima.}
\item{estadocivil}{estado civil da vitima.}
\item{profissao}{profissao da vitima.}
\item{grauinstrucao}{grau de instrucao da vitima.}
\item{corcutis}{etinia da vitima.}
\item{tipovinculo}{tipo vinculo do crime.}
\item{relacionamento}{relacionamento com o autor do crime.}
}
}
\source{
partes_tidy
}
\usage{
partes_tidy
}
\description{
Um conjunto de dados que contem informações das partes envonvidas nos crimes.
}
\keyword{datasets}
